+++ libraries/SysAPI/include/sys.hpp
@@ -8,6 +8,11 @@ namespace sys {}
 #include "sys/Cli.hpp"
 #include "sys/System.hpp"
 #include "sys/Version.hpp"
+#include "sys/TemporaryDirectory.hpp"
+
+#if defined __link
+#include "sys/Process.hpp"
+#endif
 
 using namespace sys;