+++ libraries/SysAPI/include/sys/Pipe.hpp
@@ -1,0 +1,35 @@
+#ifndef PROCESSAPI_PIPE_HPP
+#define PROCESSAPI_PIPE_HPP
+
+#if defined __link
+
+#include "fs/File.hpp"
+
+namespace sys {
+
+class Pipe : public api::ExecutionContext {
+public:
+  Pipe();
+  Pipe(const Pipe &) = delete;
+  Pipe &operator=(const Pipe &) = delete;
+  Pipe(Pipe &&) = default;
+  Pipe &operator=(Pipe &&) = default;
+
+  fs::File &read_file() { return m_read_file; }
+
+  const fs::File &read_file() const { return m_read_file; }
+
+  fs::File &write_file() { return m_write_file; }
+
+  const fs::File &write_file() const { return m_write_file; }
+
+private:
+  fs::File m_read_file;
+  fs::File m_write_file;
+};
+
+} // namespace sys
+
+#endif
+
+#endif // PROCESSAPI_PIPE_HPP