+++ libraries/SysAPI/include/sys/TemporaryDirectory.hpp
@@ -1,0 +1,20 @@
+#ifndef TEMPORARYDIRECTORY_HPP
+#define TEMPORARYDIRECTORY_HPP
+
+#include "var/StringView.hpp"
+#include "var/StackString.hpp"
+
+namespace sys {
+
+class TemporaryDirectory : public api::ExecutionContext {
+public:
+  explicit TemporaryDirectory(const var::StringView parent = "");
+  ~TemporaryDirectory();
+
+private:
+  API_AC(TemporaryDirectory, var::PathString, path);
+};
+
+} // namespace sys
+
+#endif // TEMPORARYDIRECTORY_HPP